+++ pyhidra/gui.py
@@ -11,8 +11,6 @@ logger = logging.getLogger(__name__)
 
 
 def _gui():
-    logging.basicConfig(level=logging.INFO)
-
     if platform.system() == 'Windows':
         # gui_script works like it is supposed to on windows
         gui()